refactor: extract MenuSessionLibraryCardView from MenuBarView
Move the session-library card (open-library button + pending-transcription
retry controls) and its helpers (pendingTranscriptionSummary,
openPendingTranscriptionSession, retryLatestPendingTranscription) out of
MenuBarView into a focused MenuSessionLibraryCardView under
Sources/BugNarrator/Views/Menu/. Behavior-preserving; every action still
delegates to AppState.

Second section split for #433 (#459).
