+++ Sources/BugNarrator/Views/MenuBarView.swift
@@ -32,7 +32,7 @@ struct MenuBarView: View {
             controlsSection
 
             if !transcriptStore.libraryEntries.isEmpty {
-                sessionLibraryCard
+                MenuSessionLibraryCardView(appState: appState, transcriptStore: transcriptStore)
             }
 
             MenuProductInfoView(appState: appState, isOptionKeyPressed: isOptionKeyPressed)
@@ -589,75 +589,6 @@ struct MenuBarView: View {
         }
     }
 
-    private var sessionLibraryCard: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Button("Open Session Library") {
-                appState.openTranscriptHistory()
-            }
-            .buttonStyle(.borderedProminent)
-            .controlSize(.large)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .accessibilityHint("Opens the session library window.")
-
-            if transcriptStore.pendingTranscriptionSessionCount > 0 {
-                Divider()
-
-                VStack(alignment: .leading, spacing: 8) {
-                    Text(pendingTranscriptionSummary)
-                        .font(.footnote.weight(.semibold))
-
-                    Text(
-                        appState.settingsStore.aiProvider.requiresAPIKey
-                            ? "Restore or replace the \(appState.settingsStore.aiProvider.displayName) API key in Settings if needed, then retry the saved session."
-                            : "Confirm the \(appState.settingsStore.aiProvider.displayName) setup in Settings if needed, then retry the saved session."
-                    )
-                        .font(.footnote)
-                        .foregroundStyle(.secondary)
-
-                    HStack(spacing: 10) {
-                        if appState.needsAPIKeySetup {
-                            Button("Open Settings") {
-                                appState.openSettings()
-                            }
-                            .buttonStyle(.bordered)
-                            .controlSize(.small)
-                        } else if transcriptStore.pendingTranscriptionSessionCount == 1 {
-                            Button("Retry Transcription") {
-                                retryLatestPendingTranscription()
-                            }
-                            .buttonStyle(.borderedProminent)
-                            .controlSize(.small)
-                            .disabled(appState.retryingSessionID != nil)
-                            .help("Retry transcription for the saved session without opening the library.")
-                        } else {
-                            Button("Retry Latest") {
-                                retryLatestPendingTranscription()
-                            }
-                            .buttonStyle(.bordered)
-                            .controlSize(.small)
-                            .disabled(appState.retryingSessionID != nil)
-                            .help("Retry transcription for the most recent saved session.")
-
-                            Button("Open Retry Needed Session") {
-                                openPendingTranscriptionSession()
-                            }
-                            .buttonStyle(.bordered)
-                            .controlSize(.small)
-                        }
-
-                        Button("View Library") {
-                            appState.openTranscriptHistory()
-                        }
-                        .buttonStyle(.bordered)
-                        .controlSize(.small)
-                    }
-                }
-            }
-        }
-        .padding(14)
-        .background(.quaternary.opacity(0.45), in: RoundedRectangle(cornerRadius: 14, style: .continuous))
-    }
-
     private var footerSection: some View {
         HStack(spacing: 10) {
             Button("Settings") {
@@ -724,31 +655,6 @@ struct MenuBarView: View {
             return "Use the control window to start the next session when you are ready."
         case .error:
             return "Fix the current issue, then continue from the control window."
-        }
-    }
-
-    private var pendingTranscriptionSummary: String {
-        let count = transcriptStore.pendingTranscriptionSessionCount
-        return count == 1
-            ? "1 saved session is waiting for transcription retry."
-            : "\(count) saved sessions are waiting for transcription retry."
-    }
-
-    private func openPendingTranscriptionSession() {
-        if let sessionID = transcriptStore.latestPendingTranscriptionSession?.id {
-            appState.selectedTranscriptID = sessionID
-        }
-
-        appState.openTranscriptHistory()
-    }
-
-    private func retryLatestPendingTranscription() {
-        guard let sessionID = transcriptStore.latestPendingTranscriptionSession?.id else {
-            return
-        }
-
-        Task {
-            await appState.retryPendingTranscription(for: sessionID)
         }
     }
